requirements.txt
absl-py==0.9.0
aniso8601==8.0.0
asn1crypto==1.2.0
astor==0.8.1
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.8
cycler==0.10.0
Flask==1.1.1
Flask-Cors==3.0.8
flask-restplus==0.13.0
gast==0.2.2
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.8
importlib-metadata==1.3.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
more-itertools==8.0.2
numpy==1.18.0
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.13.0
tensorboard==2.1.0
tensorflow-cpu==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.25.8
virtualenv==16.7.8
virtualenv-clone==0.5.3
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
zipp==0.6.0
